Refactor code to be easier to understand for others
Moved "utility" methods such as debit/credit an amount and methods to roundUp and isOddSplit

Moved all the type declaration to its own file

Improve test coverage
